issue #192 : Isolate failures storing location information
Isolates failures storing location information so that the rest of the record can still be updated when the service fails.

Verbosely debugs any of these errors that are not stopping records from being processed so they can be identified in logs, without having unobfuscated records from not having the SDS rules applied.

Signed-off-by: Peter Ansell <p_ansell@yahoo.com>