issue #212 : Fix linesTerminatedBy
It erroneously referred to \r\n, and did so using the unescaped character versions, interpolated by the string creation to mean CR and LF characters.

CSVWriter however defaults to LF (\n), and hence it was incorrect. Fixed to use the escaped character sequence, as directed by the Darwin Core Text guide, to be the slash character '\', and the character 'n'

Signed-off-by: Peter Ansell <p_ansell@yahoo.com>