Parsing date range end date: leap years
This is just a patch to accommodate a valid range like  2008-01-01/02-29.

Note, as a side-effect, it changes the structure of endDateParsed from MM-dd to yyyy-MM-dd, which means a processed eventdateend is written to Cassandra instead of null. I'm not sure if this should be more generally applied to the other parsers (e.g. ISODayDateRange) to set endDateParsed to a full date, where possible, instead of just the range fragment. It seems reasonable to me, but I don't have sight of the bigger picture necessarily!