Fix read of source file columns
The original values give 'array index out of bounds' errors. The source file (e.g. dd_data_NBNSYS0000188576.txt) has only 17 columns, of which the last two seem to be what is desired here.

See also the definition of  fieldsToExport, which is used to create the source file in the first place.