fixed an issue about assertion_user_id
assertion_user_id now only will be generated when user has an outstanding assertion (status = 50005 or 50001).

This fix changed the behaviour so that as long as a user makes an assertion (no matter it's verified or outstanding), his/her id will be in the assertion_user_id of corresponding occurrence record fields.